diverge BaseViewManager for UWP/WPF
