+++ ReactWindows/ReactNative.Net46/UIManager/BaseViewManager.cs
@@ -1,20 +1,11 @@
 ﻿using Newtonsoft.Json.Linq;
 using ReactNative.UIManager.Annotations;
 using System;
-#if WINDOWS_UWP
-using Windows.UI.Xaml;
-using Windows.UI.Xaml.Automation;
-using Windows.UI.Xaml.Automation.Peers;
-using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Media;
-using Windows.UI.Xaml.Media.Media3D;
-#else
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Media;
 using System.Windows.Media.Media3D;
 using System.Windows.Automation;
-#endif
 
 namespace ReactNative.UIManager
 {
@@ -67,16 +58,7 @@ namespace ReactNative.UIManager
         [ReactProp("overflow")]
         public void SetOverflow(TFrameworkElement view, string overflow)
         {
-#if WINDOWS_UWP
-            if (overflow == "hidden")
-            {
-                WinRTXamlToolkit.Controls.Extensions.FrameworkElementExtensions.SetClipToBounds(view, true);
-            }
-            else
-            {
-                WinRTXamlToolkit.Controls.Extensions.FrameworkElementExtensions.SetClipToBounds(view, false);
-            }
-#endif
+            // ToDo: Figure out how to manually manage overflow
         }
 
         /// <summary>
@@ -109,20 +91,7 @@ namespace ReactNative.UIManager
         [ReactProp("accessibilityLiveRegion")]
         public void SetAccessibilityLiveRegion(TFrameworkElement view, string liveRegion)
         {
-#if WINDOWS_UWP
-            var liveSetting = AutomationLiveSetting.Off;
-            switch (liveRegion)
-            {
-                case "polite":
-                    liveSetting = AutomationLiveSetting.Polite;
-                    break;
-                case "assertive":
-                    liveSetting = AutomationLiveSetting.Assertive;
-                    break;
-            }
-
-            AutomationProperties.SetLiveSetting(view, liveSetting);
-#endif
+            // ToDo: Figure out the WPF way of marking a region as live
         }
 
         /// <summary>
@@ -138,60 +107,12 @@ namespace ReactNative.UIManager
 
         private static void SetProjectionMatrix(TFrameworkElement view, JArray transforms)
         {
-#if WINDOWS_UWP
-            var projection = EnsureProjection(view);
-            var transformMatrix = TransformHelper.ProcessTransform(transforms);
-
-            var translateMatrix = Matrix3D.Identity;
-            var translateBackMatrix = Matrix3D.Identity;
-            if (!double.IsNaN(view.Width))
-            {
-                translateMatrix.OffsetX = -view.Width / 2;
-                translateBackMatrix.OffsetX = view.Width / 2;
-            }
-
-            if (!double.IsNaN(view.Height))
-            {
-                translateMatrix.OffsetY = -view.Height / 2;
-                translateBackMatrix.OffsetY = view.Height / 2;
-            }
-
-            projection.ProjectionMatrix = translateMatrix * transformMatrix * translateBackMatrix;
-#endif
+            // ToDo: Figure out the WPF way to do transforms
         }
 
         private static void ResetProjectionMatrix(TFrameworkElement view)
         {
-#if WINDOWS_UWP
-            var projection = view.Projection;
-            var matrixProjection = projection as Matrix3DProjection;
-            if (projection != null && matrixProjection == null)
-            {
-                throw new InvalidOperationException("Unknown projection set on framework element.");
-            }
-
-            view.Projection = null;
-#endif
+            // ToDo: Figure out the WPF way to reset transforms
         }
-
-#if WINDOWS_UWP
-        private static Matrix3DProjection EnsureProjection(FrameworkElement view)
-        {
-            var projection = view.Projection;
-            var matrixProjection = projection as Matrix3DProjection;
-            if (projection != null && matrixProjection == null)
-            {
-                throw new InvalidOperationException("Unknown projection set on framework element.");
-            }
-
-            if (matrixProjection == null)
-            {
-                matrixProjection = new Matrix3DProjection();
-                view.Projection = matrixProjection;
-            }
-
-            return matrixProjection;
-        }
-#endif
     }
 }
